requirements.txt
asgiref
dj-database-url
Django
django-stdimage
gunicorn
packaging
pillow
psycopg2
psycopg2-binary
python-dotenv
sqlparse
tzdata
whitenoise
cloudinary
djangorestframework
djangorestframework-simplejwt
markdown
gunicorn
whitenoise
